📦 chore(ploars_crate): Update Readme.md and main.rs
📝 chore(Readme.md): Add examples for applying functions to columns
✨ feat(main.rs): Implement square function for column transformation

In the `Readme.md` file, examples have been added to demonstrate how to apply functions to columns in a DataFrame using the Ploars crate. Two methods are shown: applying a closure and applying a predefined function.

In the `main.rs` file, a new function `square` has been implemented to square the values of a column. This function is then applied to the "column1" of a DataFrame in the `main` function.

Changes made:
- Added examples in the `Readme.md` file to showcase applying functions to columns using closures and predefined functions.
- Implemented the `square` function in the `main.rs` file to square the values of a column.
- Applied the `square` function to the "column1" of a DataFrame in the `main` function.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>